requirements.txt
Flask
pdf2docx
pytesseract
pdf2image
python-docx
pillow
openai-whisper==20231117